+++ packages/client/src/editorApp/editorApp.ts
@@ -63,8 +63,9 @@ export class EditorApp {
             automaticLayout: userAppConfig?.overrideAutomaticLayout ?? true
         };
         this.config.languageDef = userAppConfig?.languageDef;
+        this.config.logLevel = userAppConfig?.logLevel ?? LogLevel.Off;
 
-        this.logger.setLevel(this.config.logLevel ?? LogLevel.Off);
+        this.logger.setLevel(this.config.logLevel);
     }
 
     isDiffEditor() {
@@ -88,10 +89,6 @@ export class EditorApp {
             modified: this.modelRefs.modified?.object.textEditorModel ?? undefined,
             original: this.modelRefs.original?.object.textEditorModel ?? undefined
         };
-    }
-
-    getLogger() {
-        return this.logger;
     }
 
     registerOnTextChangedCallback(onTextChanged?: (textChanges: TextContents) => void) {
@@ -197,7 +194,7 @@ export class EditorApp {
             uri: this.config.codeResources?.modified?.uri ?? `default-uri-modified-${this.id}`,
             enforceLanguageId: this.config.codeResources?.modified?.enforceLanguageId ?? undefined
         };
-        this.modelRefs.modified = await this.buildModelReference(modified, this.logger);
+        this.modelRefs.modified = await this.buildModelReference(modified);
 
         if (this.isDiffEditor()) {
             const original = {
@@ -205,7 +202,7 @@ export class EditorApp {
                 uri: this.config.codeResources?.original?.uri ?? `default-uri-original-${this.id}`,
                 enforceLanguageId: this.config.codeResources?.original?.enforceLanguageId ?? undefined
             };
-            this.modelRefs.original = await this.buildModelReference(original, this.logger);
+            this.modelRefs.original = await this.buildModelReference(original);
         }
 
         this.logger.info(`Starting monaco-editor (${this.id})`);
@@ -255,12 +252,12 @@ export class EditorApp {
 
         if (codeResources?.modified !== undefined && codeResources.modified.uri !== this.modelRefs.modified?.object.resource.path) {
             this.modelDisposables.modified = this.modelRefs.modified;
-            this.modelRefs.modified = await this.buildModelReference(codeResources.modified, this.logger);
+            this.modelRefs.modified = await this.buildModelReference(codeResources.modified);
             updateModified = true;
         }
         if (codeResources?.original !== undefined && codeResources.original.uri !== this.modelRefs.original?.object.resource.path) {
             this.modelDisposables.original = this.modelRefs.original;
-            this.modelRefs.original = await this.buildModelReference(codeResources.original, this.logger);
+            this.modelRefs.original = await this.buildModelReference(codeResources.original);
             updateOriginal = true;
         }
 
@@ -296,7 +293,7 @@ export class EditorApp {
         await this.disposeModelRefs();
     }
 
-    async buildModelReference(codeContent: CodeContent, logger?: Logger): Promise<IReference<ITextFileEditorModel>> {
+    async buildModelReference(codeContent: CodeContent): Promise<IReference<ITextFileEditorModel>> {
         const code = codeContent.text;
         const modelRef = await createModelReference(vscode.Uri.parse(codeContent.uri), code);
 
@@ -307,7 +304,7 @@ export class EditorApp {
         const enforceLanguageId = codeContent.enforceLanguageId;
         if (enforceLanguageId !== undefined) {
             modelRef.object.setLanguageId(enforceLanguageId);
-            logger?.info(`Main languageId is enforced: ${enforceLanguageId}`);
+            this.logger.info(`Main languageId is enforced: ${enforceLanguageId}`);
         }
         return modelRef;
     };