+++ packages/examples/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this npm module are documented in this file.
 ## [2025.11.1] - unreleased
 
 - Updated to `monaco-languageclient@10.3.0` and `@typefox/monaco-editor-react@7.3.0`.
-- Updated all `@codingame/monaco-vscode` packages to `22.1.4`.
+- Updated all `@codingame/monaco-vscode` packages to `22.1.8`.
 
 ## [2025.10.1] - 2025-10-20