+++ packages/wrapper-react/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 All notable changes to npm module [@typefox/monaco-editor-react](https://www.npmjs.com/package/@typefox/monaco-editor-react) are documented in this file.
 
-## [7.3.0-next.2] - 2025-11-05
+## [7.3.0-next.3] - 2025-11-05
 
 - Fixed re-rendering issues with React StrictMode [#986](https://github.com/TypeFox/monaco-languageclient/pull/986).
-- Updated to `monaco-languageclient@10.3.0-next.2`.
-- Updated all `@codingame/monaco-vscode` packages to `22.1.4`.
+- Updated to `monaco-languageclient@10.3.0-next.3`.
+- Updated all `@codingame/monaco-vscode` packages to `22.1.8`.
 
 ## [7.2.0] - 2025-10-20